converted the news module to use the action controllers
