+++ ACP3/Modules/ACP3/News/Controller/Admin/Index/CommentsHelperTrait.php
@@ -1,0 +1,32 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\News\Controller\Admin\Index;
+
+use ACP3\Modules\ACP3\Comments;
+
+/**
+ * Class CommentsHelperTrait
+ * @package ACP3\Modules\ACP3\News\Controller\Admin\Index
+ */
+trait CommentsHelperTrait
+{
+    /**
+     * @var \ACP3\Modules\ACP3\Comments\Helpers
+     */
+    protected $commentsHelpers;
+
+    /**
+     * @param \ACP3\Modules\ACP3\Comments\Helpers $commentsHelpers
+     *
+     * @return $this
+     */
+    public function setCommentsHelpers(Comments\Helpers $commentsHelpers)
+    {
+        $this->commentsHelpers = $commentsHelpers;
+
+        return $this;
+    }
+}
