+++ ACP3/Modules/ACP3/News/Controller/Admin/Index/Index.php
@@ -1,0 +1,84 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\News\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Categories;
+use ACP3\Modules\ACP3\Comments;
+use ACP3\Modules\ACP3\News;
+
+/**
+ * Class Index
+ * @package ACP3\Modules\ACP3\News\Controller\Admin\Index
+ */
+class Index extends Core\Modules\AdminController
+{
+    /**
+     * @var \ACP3\Modules\ACP3\News\Model\NewsRepository
+     */
+    protected $newsRepository;
+
+    /**
+     * Index constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext   $context
+     * @param \ACP3\Modules\ACP3\News\Model\NewsRepository $newsRepository
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        News\Model\NewsRepository $newsRepository)
+    {
+        parent::__construct($context);
+
+        $this->newsRepository = $newsRepository;
+    }
+
+    /**
+     * @return array
+     */
+    public function execute()
+    {
+        $news = $this->newsRepository->getAllInAcp();
+
+        /** @var Core\Helpers\DataGrid $dataGrid */
+        $dataGrid = $this->get('core.helpers.data_grid');
+        $dataGrid
+            ->setResults($news)
+            ->setRecordsPerPage($this->user->getEntriesPerPage())
+            ->setIdentifier('#acp-table')
+            ->setResourcePathDelete('admin/news/index/delete')
+            ->setResourcePathEdit('admin/news/index/edit');
+
+        $dataGrid
+            ->addColumn([
+                'label' => $this->translator->t('system', 'publication_period'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\DateColumnRenderer::class,
+                'fields' => ['start', 'end'],
+                'default_sort' => true
+            ], 30)
+            ->addColumn([
+                'label' => $this->translator->t('news', 'title'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['title'],
+            ], 20)
+            ->addColumn([
+                'label' => $this->translator->t('categories', 'category'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['cat']
+            ], 20)
+            ->addColumn([
+                'label' => $this->translator->t('system', 'id'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\IntegerColumnRenderer::class,
+                'fields' => ['id'],
+                'primary' => true
+            ], 10);
+
+        return [
+            'grid' => $dataGrid->render(),
+            'show_mass_delete_button' => count($news) > 0
+        ];
+    }
+}
