+++ ACP3/Modules/ACP3/News/Controller/Admin/Index/Settings.php
@@ -1,0 +1,101 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\News\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Categories;
+use ACP3\Modules\ACP3\Comments;
+use ACP3\Modules\ACP3\News;
+
+/**
+ * Class Settings
+ * @package ACP3\Modules\ACP3\News\Controller\Admin\Index
+ */
+class Settings extends Core\Modules\AdminController
+{
+    use CommentsHelperTrait;
+
+    /**
+     * @var \ACP3\Core\Helpers\FormToken
+     */
+    protected $formTokenHelper;
+    /**
+     * @var \ACP3\Modules\ACP3\News\Validation\AdminSettingsFormValidation
+     */
+    protected $adminSettingsFormValidation;
+
+    /**
+     * Settings constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext                     $context
+     * @param \ACP3\Core\Helpers\FormToken                                   $formTokenHelper
+     * @param \ACP3\Modules\ACP3\News\Validation\AdminSettingsFormValidation $adminSettingsFormValidation
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        Core\Helpers\FormToken $formTokenHelper,
+        News\Validation\AdminSettingsFormValidation $adminSettingsFormValidation)
+    {
+        parent::__construct($context);
+
+        $this->formTokenHelper = $formTokenHelper;
+        $this->adminSettingsFormValidation = $adminSettingsFormValidation;
+    }
+
+    /**
+     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    public function actionSettings()
+    {
+        if ($this->request->getPost()->isEmpty() === false) {
+            return $this->_settingsPost($this->request->getPost()->all());
+        }
+
+        $settings = $this->config->getSettings('news');
+
+        if ($this->modules->isActive('comments') === true) {
+            $this->view->assign('allow_comments', $this->get('core.helpers.forms')->yesNoCheckboxGenerator('comments', $settings['comments']));
+        }
+
+        $this->formTokenHelper->generateFormToken();
+
+        return [
+            'dateformat' => $this->get('core.helpers.date')->dateFormatDropdown($settings['dateformat']),
+            'readmore' => $this->get('core.helpers.forms')->yesNoCheckboxGenerator('readmore', $settings['readmore']),
+            'readmore_chars' => $this->request->getPost()->get('readmore_chars', $settings['readmore_chars']),
+            'sidebar_entries' => $this->get('core.helpers.forms')->recordsPerPage((int)$settings['sidebar'], 1, 10),
+            'category_in_breadcrumb' => $this->get('core.helpers.forms')->yesNoCheckboxGenerator('category_in_breadcrumb', $settings['category_in_breadcrumb'])
+        ];
+    }
+
+    /**
+     * @param array $formData
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    protected function _settingsPost(array $formData)
+    {
+        return $this->actionHelper->handleSettingsPostAction(function () use ($formData) {
+            $this->adminSettingsFormValidation->validate($formData);
+
+            $data = [
+                'dateformat' => Core\Functions::strEncode($formData['dateformat']),
+                'sidebar' => (int)$formData['sidebar'],
+                'readmore' => $formData['readmore'],
+                'readmore_chars' => (int)$formData['readmore_chars'],
+                'category_in_breadcrumb' => $formData['category_in_breadcrumb'],
+            ];
+
+            if ($this->commentsHelpers) {
+                $data['comments'] = $formData['comments'];
+            }
+
+            $this->formTokenHelper->unsetFormToken();
+
+            return $this->config->setSettings($data, 'news');
+        });
+    }
+}
