+++ ACP3/Modules/ACP3/News/Controller/Sidebar/Index/Latest.php
@@ -1,15 +1,18 @@
 <?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
 
-namespace ACP3\Modules\ACP3\News\Controller\Sidebar;
+namespace ACP3\Modules\ACP3\News\Controller\Sidebar\Index;
 
 use ACP3\Core;
 use ACP3\Modules\ACP3\News;
 
 /**
- * Class Index
- * @package ACP3\Modules\ACP3\News\Controller\Sidebar
+ * Class Latest
+ * @package ACP3\Modules\ACP3\News\Controller\Sidebar\Index
  */
-class Index extends Core\Modules\Controller
+class Latest extends Core\Modules\Controller
 {
     /**
      * @var Core\Date
@@ -37,28 +40,9 @@ class Index extends Core\Modules\Controller
     }
 
     /**
-     * @param int    $categoryId
-     * @param string $template
-     */
-    public function actionIndex($categoryId = 0, $template = '')
-    {
-        $settings = $this->config->getSettings('news');
-
-        if (!empty($categoryId)) {
-            $news = $this->newsRepository->getAllByCategoryId((int)$categoryId, $this->date->getCurrentDateTime(), $settings['sidebar']);
-        } else {
-            $news = $this->newsRepository->getAll($this->date->getCurrentDateTime(), $settings['sidebar']);
-        }
-        $this->view->assign('sidebar_news', $news);
-        $this->view->assign('dateformat', $settings['dateformat']);
-
-        $this->setTemplate($template !== '' ? $template : 'News/Sidebar/index.index.tpl');
-    }
-
-    /**
      * @param int $categoryId
      */
-    public function actionLatest($categoryId = 0)
+    public function execute($categoryId = 0)
     {
         $settings = $this->config->getSettings('news');
