+++ ACP3/Modules/ACP3/News/Resources/config/controllers.yml
@@ -1,0 +1,92 @@
+parameters:
+  news.controller.admin.index.create.fqdn: 'ACP3\Modules\ACP3\News\Controller\Admin\Index\Create'
+  news.controller.admin.index.delete.fqdn: 'ACP3\Modules\ACP3\News\Controller\Admin\Index\Delete'
+  news.controller.admin.index.edit.fqdn: 'ACP3\Modules\ACP3\News\Controller\Admin\Index\Edit'
+  news.controller.admin.index.index.fqdn: 'ACP3\Modules\ACP3\News\Controller\Admin\Index\Index'
+  news.controller.admin.index.settings.fqdn: 'ACP3\Modules\ACP3\News\Controller\Admin\Index\Settings'
+  news.controller.frontend.index.details.fqdn: 'ACP3\Modules\ACP3\News\Controller\Frontend\Index\Details'
+  news.controller.frontend.index.index.fqdn: 'ACP3\Modules\ACP3\News\Controller\Frontend\Index\Index'
+  news.controller.sidebar.index.index.fqdn: 'ACP3\Modules\ACP3\News\Controller\Sidebar\Index\Index'
+  news.controller.sidebar.index.latest.fqdn: 'ACP3\Modules\ACP3\News\Controller\Sidebar\Index\Latest'
+
+services:
+  news.controller.admin.index.create:
+    class: %news.controller.admin.index.create.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.date
+      - @core.helpers.formToken
+      - @news.model.newsrepository
+      - @news.validator
+      - @categories.helpers
+    calls:
+      - ['setCommentsHelpers', ['@?comments.helpers']]
+
+  news.controller.admin.index.delete:
+    class: %news.controller.admin.index.delete.fqdn%
+    arguments:
+      - @core.context.admin
+      - @news.model.newsrepository
+      - @news.cache
+    calls:
+      - ['setCommentsHelpers', ['@?comments.helpers']]
+
+  news.controller.admin.index.edit:
+    class: %news.controller.admin.index.edit.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.date
+      - @core.helpers.formToken
+      - @news.model.newsrepository
+      - @news.cache
+      - @news.validator
+      - @categories.helpers
+    calls:
+      - ['setCommentsHelpers', ['@?comments.helpers']]
+
+  news.controller.admin.index.index:
+    class: %news.controller.admin.index.index.fqdn%
+    arguments:
+      - @core.context.admin
+      - @news.model.newsrepository
+
+  news.controller.admin.index.settings:
+    class: %news.controller.admin.index.settings.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @news.validation.admin_settings_form_validation
+    calls:
+      - ['setCommentsHelpers', ['@?comments.helpers']]
+
+  news.controller.frontend.index.details:
+    class: %news.controller.frontend.index.details.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @core.date
+      - @news.model.newsrepository
+      - @news.cache
+
+  news.controller.frontend.index.index:
+    class: %news.controller.frontend.index.index.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @core.date
+      - @core.pagination
+      - @news.model.newsrepository
+      - @categories.helpers
+      - @categories.model
+
+  news.controller.sidebar.index.index:
+    class: %news.controller.sidebar.index.index.fqdn%
+    arguments:
+      - '@core.context'
+      - '@core.date'
+      - '@news.model.newsrepository'
+
+  news.controller.sidebar.index.latest:
+    class: %news.controller.sidebar.index.latest.fqdn%
+    arguments:
+      - '@core.context'
+      - '@core.date'
+      - '@news.model.newsrepository'
